# openapi.yaml
openapi: 3.1.0
info:
  title: Dejima API
  version: "0.1"
  x-api-version: 1
  description: >
    HTTP/WebSocket control surface for the Dejima daemon (`dejimad`). This is the
    same API the CLI/TUI use and the contract third-party apps build on.

    Status: Dejima is in alpha (0.x). The `v1/` path prefix is intended to be
    stable, but fields may change until `v1.0.0`; an integer `api_version`
    (advertised on `/v1/overview`) tracks the client/daemon contract separately
    from the semver tag. This spec is hand-maintained from the route table in
    `internal/api/server.go` and the request/response shapes in
    `internal/api/types.go` — those are the source of truth and the spec may
    drift slightly between releases.

    WebSocket endpoints (PTY session streams) are listed for discoverability but
    cannot be fully modeled in OpenAPI; see the per-path notes and the
    `SessionEnvelope` schema for the wire protocol.
servers:
  - url: http://{host}
    description: Daemon over the tailnet (DEJIMA_HOST, e.g. 100.84.12.7:7273)
    variables:
      host:
        default: 127.0.0.1:7273
security:
  - bearerToken: []
  - {}   # operator listeners (unix socket / tailnet) carry no per-request token
...
paths:
...
  /v1/links:
    get:
      tags: [link]
      summary: List inter-island link grants (operator)
      operationId: listLinks
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LinksResponse" }
    post:
      tags: [link]
      summary: Authorize a directional inter-island info channel (operator; deny-all default)
      operationId: grantLink
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LinkGrantRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LinkGrant" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [link]
      summary: Revoke a link grant (operator)
      operationId: revokeLink
      parameters:
        - { name: from, in: query, required: true, schema: { type: string } }
        - { name: to, in: query, required: true, schema: { type: string } }
        - { name: topic, in: query, required: true, schema: { type: string } }
      responses:
        "204": { description: Revoked }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/islands/{name}/link/actions:
    parameters: [ { $ref: "#/components/parameters/Name" } ]
    get:
      tags: [link]
      summary: List the action types this island exposes for cross-island delegation
      operationId: listExposedActions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LinkExposedResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/islands/{name}/link/actions/{action}:
    parameters:
      - { $ref: "#/components/parameters/Name" }
      - { name: action, in: path, required: true, schema: { type: string } }
    put:
      tags: [link]
      summary: Expose a named action type (operator)
      operationId: exposeAction
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LinkExposedResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [link]
      summary: Stop exposing an action type (operator)
      operationId: unexposeAction
      responses:
        "204": { description: Unexposed }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/islands/{name}/link/action:
    parameters: [ { $ref: "#/components/parameters/Name" } ]
    post:
      tags: [link]
      summary: Request a named action on another island (deny-all; per-action approval)
      description: >
        The path island is the requester (token-pinned). Deny-all: requires a link
        grant for the channel AND the destination to expose the action. A
        pre-authorized action (in the grant's Actions allowlist) executes
        immediately (201, status "executed"); otherwise it is queued for operator
        approval (202, status "pending") and a link.action-pending event fires.
        Never a free-form prompt — only named, exposed action types.
      operationId: requestAction
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LinkActionRequest" }
      responses:
        "201":
          description: Executed (pre-authorized)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LinkActionResponse" }
        "202":
          description: Queued for operator approval
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LinkActionResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403":
          description: No grant, or the action is not exposed by the destination
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/link/actions/{id}/approve:
    parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
    post:
      tags: [link]
      summary: Approve and execute a pending action (operator; owner/operator only)
      operationId: approveAction
      responses:
        "200":
          description: Executed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LinkActionResponse" }
        "404": { description: Unknown or expired (fail-closed) }
        "409": { description: The grant or exposed action changed; refused }
  /v1/link/actions/{id}/deny:
    parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
    post:
      tags: [link]
      summary: Deny a pending action (operator; owner/operator only)
      operationId: denyAction
      responses:
        "204": { description: Denied }
        "404": { description: Unknown or expired }
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: Per-island bearer token (DEJIMA_TOKEN) for the autonomy path.
  parameters:
    Name:
      name: name
      in: path
      required: true
      schema: { type: string }
      description: Island name
...
  responses:
    NotFound:
      description: Island, agent, or resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
...
    LinkGrantRequest:
      type: object
      required: [from, to, topic]
      properties:
        from: { type: string, description: Source island }
        to: { type: string, description: Destination island }
        topic: { type: string, description: Named channel }
        actions: { type: array, items: { type: string }, description: "Phase 3 pre-authorized action types on this channel; empty = info-only" }
    LinkGrant:
      type: object
      properties:
        from: { type: string }
        to: { type: string }
        topic: { type: string }
        actions: { type: array, items: { type: string }, description: "Phase 3 per-action allowlist; empty = info-only" }
        created_at: { type: string, format: date-time }
    LinksResponse:
      type: object
      properties:
        grants: { type: array, items: { $ref: "#/components/schemas/LinkGrant" } }
...
    LinkActionRequest:
      type: object
      required: [to, to_agent, topic, action]
      properties:
        to: { type: string, description: Destination island }
        to_agent: { type: string, description: Destination agent id }
        topic: { type: string, description: The granted channel }
        action: { type: string, description: Named action type the destination exposed }
        params: { type: string, description: Optional typed params }
        from_agent: { type: string }
    LinkActionResponse:
      type: object
      properties:
        status: { type: string, enum: [executed, pending] }
        pending: { type: string, description: Pending approval id when status is "pending" }
    LinkExposedResponse:
      type: object
      properties:
        island: { type: string }
        actions: { type: array, items: { type: string } }
...